Move utility code out of main.go
